Move output level widget to separate component
